reorganize project directory; correct  calculation of SHORT_KEY
