+++ broswer-application/src/metadata/ecConstants.js
