+++ broswer-application/src/utils/demo/messages.proto
